Moving Metadata JComponents.
