+++ iris-visualizer/src/main/java/cfa/vo/iris/visualizer/metadata/MetadataBrowserMainView.java
@@ -32,8 +32,7 @@ import cfa.vo.iris.visualizer.preferences.VisualizerChangeEvent;
 import cfa.vo.iris.visualizer.preferences.VisualizerCommand;
 import cfa.vo.iris.visualizer.preferences.VisualizerComponentPreferences;
 import cfa.vo.iris.visualizer.preferences.VisualizerDataModel;
-import cfa.vo.iris.visualizer.stil.IrisStarJTable;
-import cfa.vo.iris.visualizer.stil.IrisStarJTable.RowSelection;
+import cfa.vo.iris.visualizer.metadata.IrisStarJTable.RowSelection;
 import cfa.vo.iris.visualizer.stil.tables.IrisStarTable;
 import cfa.vo.iris.visualizer.stil.tables.SegmentColumnInfoMatcher;
 import cfa.vo.iris.visualizer.stil.tables.UtypeColumnInfoMatcher;
@@ -185,13 +184,13 @@ public class MetadataBrowserMainView extends javax.swing.JInternalFrame {
         dataTabsPane = new javax.swing.JTabbedPane();
         plotterMetadataPanel = new javax.swing.JPanel();
         plotterMetadataScrollPane = new javax.swing.JScrollPane();
-        plotterStarJTable = new cfa.vo.iris.visualizer.stil.IrisStarJTable();
+        plotterStarJTable = new cfa.vo.iris.visualizer.metadata.IrisStarJTable();
         pointMetadataPanel = new javax.swing.JPanel();
         pointMetadataScrollPane = new javax.swing.JScrollPane();
-        pointStarJTable = new cfa.vo.iris.visualizer.stil.IrisStarJTable();
+        pointStarJTable = new cfa.vo.iris.visualizer.metadata.IrisStarJTable();
         segmentMetadataPanel = new javax.swing.JPanel();
         segmentMetadataScrollPane = new javax.swing.JScrollPane();
-        metadataJTable1 = new cfa.vo.iris.visualizer.stil.MetadataJTable();
+        metadataJTable1 = new cfa.vo.iris.visualizer.metadata.MetadataJTable();
         jMenuBar1 = new javax.swing.JMenuBar();
         fileMenu = new javax.swing.JMenu();
         extractToSedMenuItem = new javax.swing.JMenuItem();
@@ -663,13 +662,13 @@ public class MetadataBrowserMainView extends javax.swing.JInternalFrame {
     private javax.swing.JButton invertSelectionButton;
     private javax.swing.JMenuItem invertSelectionMenuItem;
     private javax.swing.JMenuBar jMenuBar1;
-    private cfa.vo.iris.visualizer.stil.MetadataJTable metadataJTable1;
+    private cfa.vo.iris.visualizer.metadata.MetadataJTable metadataJTable1;
     private javax.swing.JPanel plotterMetadataPanel;
     private javax.swing.JScrollPane plotterMetadataScrollPane;
-    private cfa.vo.iris.visualizer.stil.IrisStarJTable plotterStarJTable;
+    private cfa.vo.iris.visualizer.metadata.IrisStarJTable plotterStarJTable;
     private javax.swing.JPanel pointMetadataPanel;
     private javax.swing.JScrollPane pointMetadataScrollPane;
-    private cfa.vo.iris.visualizer.stil.IrisStarJTable pointStarJTable;
+    private cfa.vo.iris.visualizer.metadata.IrisStarJTable pointStarJTable;
     private javax.swing.JMenuItem removeMasksMenuItem;
     private javax.swing.JMenuItem restoreSetMenuItem;
     private javax.swing.JPanel segmentMetadataPanel;
