+++ iris-visualizer/src/main/java/cfa/vo/iris/visualizer/metadata/MetadataJTable.java
@@ -1,4 +1,4 @@
-package cfa.vo.iris.visualizer.stil;
+package cfa.vo.iris.visualizer.metadata;
 
 import java.util.ArrayList;
 import java.util.LinkedHashSet;
